refactor: [M3-9389] - Move Support queries
