+++ packages/manager/src/features/Support/SupportTicketDetail/SupportTicketDetail.tsx
@@ -1,6 +1,11 @@
+import {
+  useInfiniteSupportTicketRepliesQuery,
+  useProfile,
+  useSupportTicketQuery,
+} from '@linode/queries';
 import { CircleProgress, ErrorState, Stack } from '@linode/ui';
-import { styled } from '@mui/material/styles';
 import Grid from '@mui/material/Grid2';
+import { styled } from '@mui/material/styles';
 import { createLazyRoute } from '@tanstack/react-router';
 import { isEmpty } from 'ramda';
 import * as React from 'react';
@@ -9,11 +14,6 @@ import { Waypoint } from 'react-waypoint';
 
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import { LandingHeader } from 'src/components/LandingHeader';
-import { useProfile } from '@linode/queries';
-import {
-  useInfiniteSupportTicketRepliesQuery,
-  useSupportTicketQuery,
-} from 'src/queries/support';
 import { sanitizeHTML } from 'src/utilities/sanitizeHTML';
 
 import { ExpandableTicketPanel } from '../ExpandableTicketPanel';
