+++ packages/manager/src/features/Support/SupportTickets/TicketList.test.tsx
@@ -3,10 +3,12 @@ import * as React from 'react';
 
 import { supportTicketFactory } from 'src/factories';
 import { makeResourcePage } from 'src/mocks/serverHandlers';
-import { http, HttpResponse, server } from 'src/mocks/testServer';
+import { HttpResponse, http, server } from 'src/mocks/testServer';
 import { mockMatchMedia, renderWithTheme } from 'src/utilities/testHelpers';
 
-import { Props, TicketList } from './TicketList';
+import { TicketList } from './TicketList';
+
+import type { Props } from './TicketList';
 
 beforeAll(() => mockMatchMedia());
