+++ packages/manager/src/features/Support/SupportTickets/TicketList.tsx
@@ -1,3 +1,4 @@
+import { useSupportTicketsQuery } from '@linode/queries';
 import * as React from 'react';
 
 import { Hidden } from 'src/components/Hidden';
@@ -13,7 +14,6 @@ import { TableRowLoading } from 'src/components/TableRowLoading/TableRowLoading'
 import { TableSortCell } from 'src/components/TableSortCell';
 import { useOrder } from 'src/hooks/useOrder';
 import { usePagination } from 'src/hooks/usePagination';
-import { useSupportTicketsQuery } from 'src/queries/support';
 
 import { TicketRow } from './TicketRow';
 import { getStatusFilter, useTicketSeverityCapability } from './ticketUtils';
@@ -187,5 +187,3 @@ export const TicketList = (props: Props) => {
     </>
   );
 };
-
-export default TicketList;
