+++ packages/queries/src/index.ts
@@ -8,6 +8,7 @@ export * from './nodebalancers';
 export * from './placementGroups';
 export * from './profile';
 export * from './regions';
+export * from './support';
 export * from './vlans';
 export * from './vpcs';
 export * from './volumes';
